1)updated data type for digestlen

+++ ex_cli_applications/trustm_ecc_verify.c
@@ -92,7 +92,7 @@ int main (int argc, char **argv)
     uint8_t signature [150];     //To store the signture generated
     uint16_t signatureLen = sizeof(signature);
     uint8_t digest[64];
-    unsigned int digestLen=0;
+    uint16_t digestLen = 0;
     uint8_t pubkey[2048];
     uint32_t pubkeyLen;
     uint16_t pubkeySize;
